requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
flash==1.0.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
is-disposable-email==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
wtforms-validators==1.0.0
